Update dbhelper.py
Add ownership for every item in a list.
Add index to improve the performance in the get_items() and delete_item() methods.

+++ To-do Bot/dbhelper.py
@@ -8,25 +8,30 @@ class DBHelper:
 
 	def setup(self):
 		"""Creates a new table called items in the database. This table has one column (called description)"""
-		stmt = "CREATE TABLE IF NOT EXISTS items (description text)"
-		self.conn.execute(stmt)
+		tblstmt = "CREATE TABLE IF NOT EXISTS items (description text, owner text)"
+		itemidx = "CREATE INDEX IF NOT EXISTS itemIndex ON items (description ASC)"
+		ownidx = "CREATE INDEX IF NOT EXISTS ownIndex ON items (owner ASC)"
+		self.conn.execute(tblstmt)
+		self.conn.execute(itemidx)
+		self.conn.execute(ownidx)
 		self.conn.commit()
-	
-	def add_item(self, item_text):
+
+	def add_item(self, item_text, owner):
 		"""Takes the text for the item and inserts it into the database table."""
-		stmt = "INSERT INTO items (description) VALUES (?)"
-		args = (item_text, )
+		stmt = "INSERT INTO items (description, owner) VALUES (?, ?)"
+		args = (item_text, owner)
 		self.conn.execute(stmt, args)
 		self.conn.commit()
 
-	def delete_item(self, item_text):
+	def delete_item(self, item_text, owner):
 		"""Takes the text for an item and removes it from the database"""
-		stmt = "DELETE FROM items WHERE description = (?)"
-		args = (item_text, )
+		stmt = "DELETE FROM items WHERE description = (?) AND owner = (?)"
+		args = (item_text, owner)
 		self.conn.execute(stmt, args)
 		self.conn.commit()
 
-	def get_items(self):
+	def get_items(self, owner):
 		"""Returns a list of all the items in the database"""
-		stmt = "SELECT description FROM items"
-		return [x[0] for x in self.conn.execute(stmt)]
+		stmt = "SELECT description FROM items WHERE owner = (?)"
+		args = (owner, )
+		return [x[0] for x in self.conn.execute(stmt, args)]
